# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.CallActionRequest:
    properties:
      action:
        example: answered
        type: string
      call_id:
        example: mqtt-call-20250510-abcdef123456
        type: string
      reason:
        example: user_busy
        type: string
      timestamp:
        example: 1651234567890
        type: integer
    required:
    - action
    - call_id
    type: object
...
  controllers.CallInfo:
    properties:
      action:
        example: answered
        type: string
      call_id:
        example: mqtt-call-20250510-abcdef123456
        type: string
      reason:
        example: user_busy
        type: string
      timestamp:
        example: 1651234567890
        type: integer
    type: object
...
  controllers.CallSessionResponse:
    properties:
      call_id:
        example: mqtt-call-20250510-abcdef123456
        type: string
      device_id:
        example: "5"
        type: string
      last_activity:
        example: "2025-05-10T15:09:10Z"
        type: string
      resident_id:
        example: "6"
        type: string
      start_time:
        example: "2025-05-10T15:04:05Z"
        type: string
      status:
        example: connected
        type: string
      tencen_rtc:
        $ref: '#/definitions/controllers.TRTCInfo'
    type: object
...
  controllers.EndCallSessionRequest:
    properties:
      call_id:
        example: mqtt-call-20250510-abcdef123456
        type: string
      reason:
        example: call_completed
        type: string
    required:
    - call_id
    type: object
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
  controllers.InitiateCallRequest:
    properties:
      device_id:
        description: 设备ID
        example: "5"
        type: string
      household_number:
        description: 可选，指定户号
        example: MQTT-101
        type: string
      timestamp:
        description: 可选时间戳
        example: 1651234567890
        type: integer
    required:
    - device_id
    type: object
  controllers.InitiateCallResponse:
    properties:
      call_id:
        example: mqtt-call-20250510-abcdef123456
        type: string
      call_info:
        $ref: '#/definitions/controllers.CallInfo'
      device_device_id:
        example: "5"
        type: string
      target_resident_ids:
        example:
        - '["6"'
        - '"7"]'
        items:
          type: string
        type: array
      tencen_rtc:
        $ref: '#/definitions/controllers.TRTCInfo'
      timestamp:
        example: 1651234567890
        type: integer
    type: object
...
  controllers.PublishDeviceStatusRequest:
    properties:
      battery:
        example: 85
        type: integer
      device_id:
        example: "5"
        type: string
      online:
        example: true
        type: boolean
      properties:
        additionalProperties: true
        type: object
    required:
    - device_id
    type: object
  controllers.PublishSystemMessageRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      level:
        example: info
        type: string
      message:
        example: 系统将于今晚22:00进行升级维护
        type: string
      timestamp:
        example: 1651234567890
        type: integer
      type:
        example: notification
        type: string
    required:
    - level
    - message
    - type
    type: object
...
  controllers.TRTCInfo:
    properties:
      room_id:
        example: room_5_6_1746870072
        type: string
      room_id_type:
        example: string
        type: string
      sdk_app_id:
        example: 1400000001
        type: integer
      user_id:
        example: "5"
        type: string
      user_sig:
        example: eJwtzM1Og0AUhmG...
        type: string
    type: object
...
host: 39.108.49.167:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: intercom_http_service API
  version: "1.0"
paths:
...
  /mqtt/call:
    post:
      consumes:
      - application/json
      description: 通过MQTT向设备关联的住户发起视频通话请求。如果提供了household_number参数，则呼叫该户号下的所有住户；如果未提供，则呼叫该设备绑定的户号下的所有住户。
      parameters:
      - description: 通话请求参数：支持device_id(必填)、household_number(可选)、timestamp(可选)参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.InitiateCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.InitiateCallResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 发起MQTT通话
      tags:
      - MQTT
  /mqtt/controller/device:
    post:
      consumes:
      - application/json
      description: 处理设备端通话动作，支持的动作类型包括：hangup(挂断)、cancelled(取消呼叫)
      parameters:
      - description: 设备通话动作请求，包含call_id和action字段
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 处理MQTT呼叫方动作
      tags:
      - MQTT
  /mqtt/controller/resident:
    post:
      consumes:
      - application/json
      description: 处理居民端通话动作，支持的动作类型包括：rejected(拒绝)、answered(接听)、hangup(挂断)、timeout(超时)
      parameters:
      - description: 居民通话动作请求，包含call_id、action和可选的reason字段
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 处理MQTT被呼叫方动作
      tags:
      - MQTT
  /mqtt/device/status:
    post:
      consumes:
      - application/json
      description: 更新设备状态信息，包括在线状态、电池电量和其他自定义属性，无需MQTT连接，系统会通过MQTT推送给相关订阅方
      parameters:
      - description: 设备状态信息：必须包含device_id、online、battery字段，可选包含properties自定义属性
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.PublishDeviceStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 更新设备状态
      tags:
      - Device
  /mqtt/end-session:
    post:
      consumes:
      - application/json
      description: 强制结束通话会话并通知所有参与方，适用于系统管理或异常情况下的通话强制终止
      parameters:
      - description: 结束通话会话请求，包含call_id和可选的reason字段
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EndCallSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 结束MQTT通话会话
      tags:
      - MQTT
  /mqtt/session:
    get:
      consumes:
      - application/json
      description: 获取通话会话信息及TRTC房间详情，包括设备ID、住户ID、通话状态、开始时间等
      parameters:
      - description: 通话会话ID
        in: query
        name: call_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CallSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 获取MQTT通话会话
      tags:
      - MQTT
  /mqtt/system/message:
    post:
      consumes:
      - application/json
      description: 通过MQTT发布系统消息，支持info、warning、error三种级别，消息会推送给所有订阅相关主题的客户端
      parameters:
      - description: 系统消息信息：必须包含type、level、message字段，level支持info、warning、error三种级别
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.PublishSystemMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 发布系统消息
      tags:
      - MQTT
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"